+++ cache-config/t3c-apply/config/config.go
@@ -89,7 +89,7 @@ type Cfg struct {
 	TOPass              string
 	TOURL               string
 	DNSLocalBind        bool
-	WaitForParents      bool
+	WaitForParents      WaitForParentsFlag
 	YumOptions          string
 	// UseGit is whether to create and maintain a git repo of config changes.
 	// Note this only applies to the ATS config directory inferred or set via the flag.
@@ -127,6 +127,31 @@ func StrToUseGitFlag(str string) UseGitFlag {
 		return UseGitFlag(str)
 	default:
 		return UseGitInvalid
+	}
+}
+
+type WaitForParentsFlag string
+
+const WaitForParentsDefault = WaitForParentsReval
+
+const (
+	WaitForParentsTrue    = "true"
+	WaitForParentsFalse   = "false"
+	WaitForParentsReval   = "reval"
+	WaitForParentsInvalid = ""
+)
+
+func StrToWaitForParentsFlag(str string) WaitForParentsFlag {
+	str = strings.ToLower(strings.TrimSpace(str))
+	switch str {
+	case WaitForParentsTrue:
+		fallthrough
+	case WaitForParentsFalse:
+		fallthrough
+	case WaitForParentsReval:
+		return WaitForParentsFlag(str)
+	default:
+		return WaitForParentsInvalid
 	}
 }
 
@@ -202,7 +227,7 @@ func GetCfg() (Cfg, error) {
 	toUserPtr := getopt.StringLong("traffic-ops-user", 'U', "", "Traffic Ops username. Required. May also be set with the environment variable TO_USER")
 	toPassPtr := getopt.StringLong("traffic-ops-password", 'P', "", "Traffic Ops password. Required. May also be set with the environment variable TO_PASS")
 	tsHomePtr := getopt.StringLong("trafficserver-home", 'R', "", "Trafficserver Package directory. May also be set with the environment variable TS_HOME")
-	waitForParentsPtr := getopt.BoolLong("wait-for-parents", 'W', "[true | false] do not update if parent_pending = 1 in the update json. default is false, wait for parents")
+	waitForParentsPtr := getopt.StringLong("wait-for-parents", 'W', "reval", "[true | false | reval] do not update if parent_pending = 1 in the update json. default is reval, wait for parents in reval mode but not syncds (unless Traffic Ops !UseRevalPending, and then wait in syncds)")
 	dnsLocalBindPtr := getopt.BoolLong("dns-local-bind", 'b', "[true | false] whether to use the server's Service Addresses to set the ATS DNS local bind address")
 	helpPtr := getopt.BoolLong("help", 'h', "Print usage information and exit")
 	useGitStr := getopt.StringLong("git", 'g', "auto", "Create and use a git repo in the config directory. Options are yes, no, and auto. If yes, create and use. If auto, use if it exist. Default is auto.")
@@ -238,6 +263,11 @@ func GetCfg() (Cfg, error) {
 		return Cfg{}, errors.New("Invalid git flag '" + *useGitStr + "'. Valid options are yes, no, auto.")
 	}
 
+	waitForParents := StrToWaitForParentsFlag(*waitForParentsPtr)
+	if waitForParents == WaitForParentsInvalid {
+		return Cfg{}, errors.New("Invalid wait-for-parents flag '" + *waitForParentsPtr + "'. Valid options are true, false, reval.")
+	}
+
 	retries := *retriesPtr
 	revalWaitTime := time.Second * time.Duration(*revalWaitTimePtr)
 	reverseProxyDisable := *reverseProxyDisablePtr
@@ -247,7 +277,6 @@ func GetCfg() (Cfg, error) {
 	toURL := *toURLPtr
 	toUser := *toUserPtr
 	toPass := *toPassPtr
-	waitForParents := *waitForParentsPtr
 	dnsLocalBind := *dnsLocalBindPtr
 	help := *helpPtr
 	maxmindLocation := *maxmindLocationPtr
@@ -430,7 +459,7 @@ func printConfig(cfg Cfg) {
 	log.Debugf("TOPass: Pass len: '%d'\n", len(cfg.TOPass))
 	log.Debugf("TOURL: %s\n", cfg.TOURL)
 	log.Debugf("TSHome: %s\n", TSHome)
-	log.Debugf("WaitForParents: %t\n", cfg.WaitForParents)
+	log.Debugf("WaitForParents: %v\n", cfg.WaitForParents)
 	log.Debugf("YumOptions: %s\n", cfg.YumOptions)
 	log.Debugf("MaxmindLocation: %s\n", cfg.MaxMindLocation)
 }